⭐️ Impl. `shouldUseDiscoverabilityTitleAsFallbackValueForSubtitle`
Summary: On iOS 15+, the `MenuActionConfig.discoverabilityTitle` is no longer displayed as as a subtitle under the menu action (instead you need to set a different property called `subtitle`).

This prop will revert the old behavior for backwards compatibility.

The `discoverabilityTitle` property is now used for the "discoverability heads-up display" (e.g when an app supports keyboard shortcuts, holding down the command key presents a list of shortcuts; the `discoverabilityTitle` is then used as the title for the shortcut).