Fix accessibility issues in Dropdown Menus
To VoiceOver, the dropdown menu trigger button is just read out as
'button'.

This is because by default, a UIButton has isAccessibilityElement=true.
This means that everything inside of it is invisible to the screen
reader.

In order to improve accessibility, it's necessary to set
isAccessibilityElement to false so that the screen reader can access the
react native component that the user has provided to be the trigger
button for the context menu.